+++ api/index.js
@@ -4,19 +4,21 @@ import cors from "cors";
 import mongo from "mongodb";
 import mongoose from "mongoose";
 import mongoSanitize from "express-mongo-sanitize";
-// import nodemailer from "nodemailer";
 import { clerkClient } from "@clerk/express";
 
+// util functions
+import { validateInput } from "./utils/validate-utils.js";
+import { deleteLevelTranslations, createLevelTranslations } from "./utils/translation-utils.js";
+
 // external schemas
-import Translation from "./schemas/Translation.js";
 import User from "./schemas/User.js";
-// import Contact from "./schemas/Contact.js";
 import Level from "./schemas/Level.js";
 import { Class, Conversation } from './schemas/Classes.js';
 
 // external routes
 import translationRoutes from './routes/translation-routes.js';
 import emailRoutes from './routes/email-routes.js';
+import userRoutes from './routes/user-routes.js';
 
 const app = express()
 app.use(cors())
@@ -25,6 +27,7 @@ app.use(mongoSanitize())
 
 app.use('/api/locales', translationRoutes);
 app.use('/api/', emailRoutes);
+app.use('/api/', userRoutes);
 
 const PORT = process.env.PORT || 4000;
 mongoose.connect(process.env.MONGODB_URI)
@@ -55,221 +58,69 @@ app.get('/', (req, res) => {
 });
 
 
-
-//------------------ HELPER FUNCTIONS ------------------//
-
-/*
-purpose: check that the input key is allowed
-argument types:
-  inputs: object
-  allowedFields: array
-return type:
-  array containing fields that are in allowedFields
-
-example of using to get filters for classes: validateInput(req.query, classFields)
-*/
-const validateInput = (input, allowedFields) => {
-  const filteredInput = {}
-
-  for (const key in input) {
-    if (allowedFields.includes(key)) {
-      filteredInput[key] = input[key]
-    }
-  }
-
-  return filteredInput
-}
-
-const formattedSkillKey = (skill) => `level_skill_${skill.toLowerCase().replace(/ /g, "_")}`;
-
-const deleteLevelTranslations = async (levelData) => {
-  try {
-    await Translation.deleteMany({ key: `level_name_${levelData._id}` });
-    await Translation.deleteMany({ key: `level_desc_${levelData._id}` });
-    for (const skill of levelData.skills) {
-      const key = `${formattedSkillKey(skill)}_${levelData._id}`;
-      await Translation.deleteMany({ key });
-    }
-  } catch (error) {
-    console.error("Failed to delete level translations", error);
-    throw new Error("Failed to delete level translations");
-  }
-};
-
-const createLevelTranslations = async (levelData) => {
-  try {
-    // name translation
-    const response = await fetch(`https://api.i18nexus.com/project_resources/base_strings.json?api_key=${process.env.I18NEXUS_API_KEY}`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        "Authorization": `Bearer ${process.env.I18NEXUS_PAT}`
-      },
-      body: JSON.stringify({
-        key: `level_name_${levelData._id}`,
-        value: levelData.name,
-        namespace: "levels"
-      })
-    });
-    const data = await response.json();
-    if (!response.ok) {
-      console.error("Failed to create translation", data);
-    }
-    // description translation
-    await fetch(`https://api.i18nexus.com/project_resources/base_strings.json?api_key=${process.env.I18NEXUS_API_KEY}`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        "Authorization": `Bearer ${process.env.I18NEXUS_PAT}`
-      },
-      body: JSON.stringify({
-        key: `level_desc_${levelData._id}`,
-        value: levelData.description,
-        namespace: "levels"
-      })
-    });
-    // skill translations
-    for (const skill of levelData.skills) {
-      const key = `${formattedSkillKey(skill)}_${levelData._id}`;
-      await fetch(`https://api.i18nexus.com/project_resources/base_strings.json?api_key=${process.env.I18NEXUS_API_KEY}`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          "Authorization": `Bearer ${process.env.I18NEXUS_PAT}`
-        },
-        body: JSON.stringify({
-          key,
-          value: skill,
-          namespace: "levels"
-        })
-      });
-    }
-  } catch (error) {
-    throw new Error("Failed to create level translations");
-  }
-}
-
-
 //------------------ ENDPOINTS ------------------//
 
 
 /* USER RELATED ENDPOINTS */
 
-// Sign up
-app.post('/api/sign-up', async (req, res) => {
-  try {
-    const { firstName, lastName, email, whatsapp, password, clerkId } = req.body;
+// // Sign up
+// app.post('/api/sign-up', async (req, res) => {
+//   try {
+//     const { firstName, lastName, email, whatsapp, password, clerkId } = req.body;
 
-    // Check if user already exists
-    const existingUser = await User.findOne({ email });
+//     // Check if user already exists
+//     const existingUser = await User.findOne({ email });
 
-    if (existingUser) {
-      return res.status(409).json({ message: 'Email already exists' });
-    }
+//     if (existingUser) {
+//       return res.status(409).json({ message: 'Email already exists' });
+//     }
 
-    // Create new user with separate first/last name fields
-    const newUser = new User({
-      firstName,
-      lastName,
-      email,
-      whatsapp,
-      password,
-      clerkId
-    });
+//     // Create new user with separate first/last name fields
+//     const newUser = new User({
+//       firstName,
+//       lastName,
+//       email,
+//       whatsapp,
+//       password,
+//       clerkId
+//     });
 
-    await newUser.save();
-    res.status(201).json(newUser);
+//     await newUser.save();
+//     res.status(201).json(newUser);
 
-  } catch (error) {
-    console.error('Failed to sign up:', error);
-    res.status(500).json({ message: 'Failed to sign up' });
-  }
-})
+//   } catch (error) {
+//     console.error('Failed to sign up:', error);
+//     res.status(500).json({ message: 'Failed to sign up' });
+//   }
+// })
 
-// Get Users
-app.get('/api/users', async (req, res) => {
-  try {
-    const users = await User.find();
-    return res.status(200).json(users);
-  } catch (err) {
-    res.status(500).send(err);
-  }
-})
-
-
-// Get User
-app.get('/api/user', async (req, res) => {
-  const allowedFields = ['_id', 'email', 'whatsapp']
-  const filters = validateInput(req.query, allowedFields)
-
-  if (Object.keys(filters).length === 0) {
-    res.status(404).send('Error: user not found', err);
-  }
-
-  try {
-    const user = await User.findOne(filters);
-    res.status(200).json(user);
-  } catch (err) {
-    res.status(500).send(err);
-  }
-})
+// // Get Users
+// app.get('/api/users', async (req, res) => {
+//   try {
+//     const users = await User.find();
+//     return res.status(200).json(users);
+//   } catch (err) {
+//     res.status(500).send(err);
+//   }
+// })
 
 
-/* CONTACT RELATED ENDPOINTS */
+// // Get User
+// app.get('/api/user', async (req, res) => {
+//   const allowedFields = ['_id', 'email', 'whatsapp']
+//   const filters = validateInput(req.query, allowedFields)
 
-// Post Contact
-// app.post('/api/contact', async (req, res) => {
-//   const { name, email, subject, message } = req.body
+//   if (Object.keys(filters).length === 0) {
+//     res.status(404).send('Error: user not found', err);
+//   }
 
 //   try {
-//     const newContact = new Contact({
-//       name,
-//       email,
-//       subject,
-//       message
-//     });
-//     await newContact.save();
-
-//     // Nodemailer setup
-//     const transporter = nodemailer.createTransport({
-//       service: 'gmail',
-//       auth: {
-//         user: process.env.ADMIN_EMAIL,
-//         pass: process.env.ADMIN_PASSWORD,
-//       },
-//     });
-
-//     transporter.verify((error, success) => {
-//       if (error) {
-//         console.error('Error initializing transporter:', error);
-//       } else {
-//         console.log('Transporter is ready to send emails', success);
-//       }
-//     });
-
-
-//     const mailOptions = {
-//       from: email,
-//       to: process.env.ADMIN_EMAIL,
-//       subject: `Contact Form: ${subject}`,
-//       html: `
-//         <p><strong>From:</strong> ${name} (${email})</p>
-//         <p><strong>Subject:</strong> ${subject}</p>
-//         <p><strong>Message:</strong></p>
-//         <p>${message}</p>
-//       `
-//     };
-
-//     // Send email
-//     await transporter.sendMail(mailOptions);
-
-//     res.status(201).json({ message: 'Inquiry and email submitted successfully' });
+//     const user = await User.findOne(filters);
+//     res.status(200).json(user);
+//   } catch (err) {
+//     res.status(500).send(err);
 //   }
-//   catch (err) {
-//     console.error('Error submitting inquiry:', err);
-//     res.status(500).json({ message: 'Error submitting inquiry', error: err.message });
-//   }
-// });
+// })
 
 
 /* CLASS RELATED ENDPOINTS */
