+++ src/pages/dashboards/admin/AdminStudents.jsx
@@ -4,7 +4,8 @@ import { useLocation, Link } from 'wouter';
 import { useAuth } from '@clerk/clerk-react';
 import { IoPersonOutline } from "react-icons/io5";
 import { getUsers, getStudentsForExport } from '@/api/user-wrapper.js';
-import { getClasses, getStudentsClasses, getClassById } from '@/api/class-wrapper';
+import { getClasses, getClassById } from '@/api/class-wrapper';
+import { getStudentsClasses } from '@/api/user-wrapper';
 import Unauthorized from "@/pages/Unauthorized";
 import Dropdown from '@/components/Dropdown/Dropdown';
 import Button from '@/components/Button/Button';
